requirements.txt
requests
pytest
pytest-cov
discord.py
